Separate editor panels into structs and start work on scene view inspector linking
Took 1 hour 38 minutes
